refactor(asc): group client queries by resource
